+++ x/globalfee/feechecker/feechecker_test.go
@@ -5,7 +5,6 @@ import (
 	"testing"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 	"github.com/stretchr/testify/suite"
 
 	"github.com/bandprotocol/chain/v2/testing/testapp"
@@ -93,8 +92,7 @@ func (suite *FeeCheckerTestSuite) TestValidRawReport() {
 	stubTx := &StubTx{Msgs: msgs}
 
 	// test - check report tx
-	isReportTx, err := suite.FeeChecker.CheckReportTx(suite.ctx, stubTx)
-	suite.Require().NoError(err)
+	isReportTx := suite.FeeChecker.CheckReportTx(suite.ctx, stubTx)
 	suite.Require().True(isReportTx)
 
 	// test - check tx fee with min gas prices
@@ -109,12 +107,11 @@ func (suite *FeeCheckerTestSuite) TestNotValidRawReport() {
 	stubTx := &StubTx{Msgs: msgs}
 
 	// test - check report tx
-	isReportTx, err := suite.FeeChecker.CheckReportTx(suite.ctx, stubTx)
-	suite.Require().Error(err)
+	isReportTx := suite.FeeChecker.CheckReportTx(suite.ctx, stubTx)
 	suite.Require().False(isReportTx)
 
 	// test - check tx fee with min gas prices
-	_, _, err = suite.FeeChecker.CheckTxFeeWithMinGasPrices(suite.ctx, stubTx)
+	_, _, err := suite.FeeChecker.CheckTxFeeWithMinGasPrices(suite.ctx, stubTx)
 	suite.Require().Error(err)
 }
 
@@ -126,8 +123,7 @@ func (suite *FeeCheckerTestSuite) TestValidReport() {
 	stubTx := &StubTx{Msgs: []sdk.Msg{&authzMsg}}
 
 	// test - check report tx
-	isReportTx, err := suite.FeeChecker.CheckReportTx(suite.ctx, stubTx)
-	suite.Require().NoError(err)
+	isReportTx := suite.FeeChecker.CheckReportTx(suite.ctx, stubTx)
 	suite.Require().True(isReportTx)
 
 	// test - check tx fee with min gas prices
@@ -142,16 +138,15 @@ func (suite *FeeCheckerTestSuite) TestNoAuthzReport() {
 		types.NewMsgReportData(suite.requestId, []types.RawReport{}, testapp.Validators[0].ValAddress),
 	}
 	authzMsg := authz.NewMsgExec(testapp.Bob.Address, reportMsgs)
-	stubTx := &StubTx{Msgs: []sdk.Msg{&authzMsg}}
+	stubTx := &StubTx{Msgs: []sdk.Msg{&authzMsg}, GasPrices: sdk.NewDecCoins(sdk.NewDecCoin("uband", sdk.NewInt(1)))}
 
 	// test - check report tx
-	isReportTx, err := suite.FeeChecker.CheckReportTx(suite.ctx, stubTx)
-	suite.Require().EqualError(err, sdkerrors.ErrUnauthorized.Wrap("authorization not found").Error())
+	isReportTx := suite.FeeChecker.CheckReportTx(suite.ctx, stubTx)
 	suite.Require().False(isReportTx)
 
 	// test - check tx fee with min gas prices
-	_, _, err = suite.FeeChecker.CheckTxFeeWithMinGasPrices(suite.ctx, stubTx)
-	suite.Require().EqualError(err, sdkerrors.ErrUnauthorized.Wrap("authorization not found").Error())
+	_, _, err := suite.FeeChecker.CheckTxFeeWithMinGasPrices(suite.ctx, stubTx)
+	suite.Require().NoError(err)
 }
 
 func (suite *FeeCheckerTestSuite) TestNotValidReport() {
@@ -162,12 +157,11 @@ func (suite *FeeCheckerTestSuite) TestNotValidReport() {
 	stubTx := &StubTx{Msgs: []sdk.Msg{&authzMsg}}
 
 	// test - check report tx
-	isReportTx, err := suite.FeeChecker.CheckReportTx(suite.ctx, stubTx)
-	suite.Require().Error(err)
+	isReportTx := suite.FeeChecker.CheckReportTx(suite.ctx, stubTx)
 	suite.Require().False(isReportTx)
 
 	// test - check tx fee with min gas prices
-	_, _, err = suite.FeeChecker.CheckTxFeeWithMinGasPrices(suite.ctx, stubTx)
+	_, _, err := suite.FeeChecker.CheckTxFeeWithMinGasPrices(suite.ctx, stubTx)
 	suite.Require().Error(err)
 }
 
@@ -183,18 +177,28 @@ func (suite *FeeCheckerTestSuite) TestNotReportMsg() {
 		testapp.TestDefaultExecuteGas,
 		testapp.FeePayer.Address,
 	)
-	stubTx := &StubTx{Msgs: []sdk.Msg{requestMsg}, GasPrices: sdk.NewDecCoins(sdk.NewDecCoin("uband", sdk.NewInt(1)))}
+	stubTx := &StubTx{
+		Msgs: []sdk.Msg{requestMsg},
+		GasPrices: sdk.NewDecCoins(
+			sdk.NewDecCoinFromDec("uaaaa", sdk.NewDecWithPrec(100, 3)),
+			sdk.NewDecCoinFromDec("uaaab", sdk.NewDecWithPrec(1, 3)),
+			sdk.NewDecCoinFromDec("uaaac", sdk.NewDecWithPrec(0, 3)),
+			sdk.NewDecCoinFromDec("uband", sdk.NewDecWithPrec(3, 3)),
+			sdk.NewDecCoinFromDec("uccca", sdk.NewDecWithPrec(0, 3)),
+			sdk.NewDecCoinFromDec("ucccb", sdk.NewDecWithPrec(1, 3)),
+			sdk.NewDecCoinFromDec("ucccc", sdk.NewDecWithPrec(100, 3)),
+		),
+	}
 
 	// test - check report tx
-	isReportTx, err := suite.FeeChecker.CheckReportTx(suite.ctx, stubTx)
-	suite.Require().NoError(err)
+	isReportTx := suite.FeeChecker.CheckReportTx(suite.ctx, stubTx)
 	suite.Require().False(isReportTx)
 
 	// test - check tx fee with min gas prices
 	fee, priority, err := suite.FeeChecker.CheckTxFeeWithMinGasPrices(suite.ctx, stubTx)
 	suite.Require().NoError(err)
 	suite.Require().Equal(stubTx.GetFee(), fee)
-	suite.Require().Equal(int64(1000), priority)
+	suite.Require().Equal(int64(30), priority)
 }
 
 func (suite *FeeCheckerTestSuite) TestReportMsgAndOthersTypeMsgInTheSameAuthzMsgs() {
@@ -215,15 +219,14 @@ func (suite *FeeCheckerTestSuite) TestReportMsgAndOthersTypeMsgInTheSameAuthzMsg
 	stubTx := &StubTx{Msgs: []sdk.Msg{&authzMsg}, GasPrices: sdk.NewDecCoins(sdk.NewDecCoin("uband", sdk.NewInt(1)))}
 
 	// test - check report tx
-	isReportTx, err := suite.FeeChecker.CheckReportTx(suite.ctx, stubTx)
-	suite.Require().NoError(err)
+	isReportTx := suite.FeeChecker.CheckReportTx(suite.ctx, stubTx)
 	suite.Require().False(isReportTx)
 
 	// test - check tx fee with min gas prices
 	fee, priority, err := suite.FeeChecker.CheckTxFeeWithMinGasPrices(suite.ctx, stubTx)
 	suite.Require().NoError(err)
 	suite.Require().Equal(stubTx.GetFee(), fee)
-	suite.Require().Equal(int64(1000), priority)
+	suite.Require().Equal(int64(10000), priority)
 }
 
 func (suite *FeeCheckerTestSuite) TestReportMsgAndOthersTypeMsgInTheSameTx() {
@@ -245,15 +248,14 @@ func (suite *FeeCheckerTestSuite) TestReportMsgAndOthersTypeMsgInTheSameTx() {
 	}
 
 	// test - check report tx
-	isReportTx, err := suite.FeeChecker.CheckReportTx(suite.ctx, stubTx)
-	suite.Require().NoError(err)
+	isReportTx := suite.FeeChecker.CheckReportTx(suite.ctx, stubTx)
 	suite.Require().False(isReportTx)
 
 	// test - check tx fee with min gas prices
 	fee, priority, err := suite.FeeChecker.CheckTxFeeWithMinGasPrices(suite.ctx, stubTx)
 	suite.Require().NoError(err)
 	suite.Require().Equal(stubTx.GetFee(), fee)
-	suite.Require().Equal(int64(1000), priority)
+	suite.Require().Equal(int64(10000), priority)
 }
 
 func (suite *FeeCheckerTestSuite) TestGetBondDenom() {