+++ x/globalfee/genesis_test.go
@@ -23,7 +23,7 @@ import (
 func TestDefaultGenesis(t *testing.T) {
 	encCfg := simapp.MakeTestEncodingConfig()
 	gotJSON := AppModuleBasic{}.DefaultGenesis(encCfg.Codec)
-	assert.JSONEq(t, `{"params":{"minimum_gas_prices":[]}}`, string(gotJSON), string(gotJSON))
+	assert.JSONEq(t, `{"params":{"minimum_gas_prices":[]}}`, string(gotJSON))
 }
 
 func TestValidateGenesis(t *testing.T) {
@@ -41,9 +41,9 @@ func TestValidateGenesis(t *testing.T) {
 		"minimum not set": {
 			src: `{"params":{}}`,
 		},
-		"zero amount allowed": {
+		"zero amount not allowed": {
 			src:    `{"params":{"minimum_gas_prices":[{"denom":"ALX", "amount":"0"}]}}`,
-			expErr: false,
+			expErr: true,
 		},
 		"duplicate denoms not allowed": {
 			src:    `{"params":{"minimum_gas_prices":[{"denom":"ALX", "amount":"1"},{"denom":"ALX", "amount":"2"}]}}`,